KAN-16 Refactor: Change Set Name and Format code
