+++ src/main/java/capstone/relation/GoodRelationApplication.java
@@ -1,4 +1,4 @@
-package good_relation.good_relation;
+package capstone.relation;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
