+++ src/main/java/capstone/relation/websocket/WebSocketConfig.java
@@ -1,4 +1,5 @@
-package good_relation.good_relation.websocket;
+package capstone.relation.websocket;
+
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
