+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=good_relation
+spring.application.name=good_relationship
