+++ src/main/resources/static/index.html
@@ -30,7 +30,9 @@
     <div>
         <table id="conversation" class="table table-striped">
             <thead>
-            <tr><th>Messages</th></tr>
+            <tr>
+                <th>Messages</th>
+            </tr>
             </thead>
             <tbody id="message"></tbody>
         </table>
